+++ brikar-common/src/main/java/com/truward/brikar/common/tracking/TrackingHttpHeaderNames.java
@@ -22,34 +22,34 @@ public final class TrackingHttpHeaderNames {
   private static final String SUFFIX_HEX_CHARS = "0123456789ABCDEF";
 
   /**
-   * Header name, that should store request ID.
+   * Header name, that should store request vector.
    * It may also include optional components separated by dots.
    * These components identify nested requests to make it easier to populate request graph.
    *
    * Once retrieved the corresponding value should be stored in {@link org.slf4j.MDC} context.
-   * See also {@link com.truward.brikar.common.log.LogUtil#REQUEST_ID}.
+   * See also {@link com.truward.brikar.common.log.LogUtil#REQUEST_VECTOR}.
    */
-  public static final String REQUEST_ID = "Request-ID";
+  public static final String REQUEST_VECTOR = "RV";
 
-  public static String getNestedRequestId(String requestId, long suffix) {
+  public static String getNestedRequestVector(String requestVector, long suffix) {
     if (suffix < 0) {
       throw new IllegalArgumentException("suffix");
     }
 
-    if (!StringUtils.hasLength(requestId)) {
-      throw new IllegalArgumentException("requestId");
+    if (!StringUtils.hasLength(requestVector)) {
+      throw new IllegalArgumentException("requestVector");
     }
 
     // allocate string builder with estimated size of the resultant request
     final int numberOfHexLeadingZeros = Long.numberOfLeadingZeros(suffix) / 4;
-    final int suffixSize = 16 - numberOfHexLeadingZeros;
+    final int suffixSize = 16 - numberOfHexLeadingZeros; // 16 is the max size of hex representation of the long number
 
-    int offset = requestId.length();
+    int offset = requestVector.length();
     final char[] result = new char[offset + 1 + suffixSize];
 
     // first, copy original request ID
     for (int i = 0; i < offset; ++i) {
-      result[i] = requestId.charAt(i);
+      result[i] = requestVector.charAt(i);
     }
 
     // then append dot
@@ -66,47 +66,47 @@ public final class TrackingHttpHeaderNames {
     return new String(result);
   }
 
-  public static String getNestedRequestId(String requestId) {
-    return getNestedRequestId(requestId, REQUEST_ID_COUNTER.incrementAndGet());
+  public static String getNestedRequestVector(String requestVector) {
+    return getNestedRequestVector(requestVector, REQUEST_ID_COUNTER.incrementAndGet());
   }
 
-  public static RequestIdParts parseRequestId(String requestId) {
-    if (!StringUtils.hasLength(requestId)) {
-      throw new IllegalArgumentException("requestId");
+  public static RequestVectorParts parseRequestVector(String requestVector) {
+    if (!StringUtils.hasLength(requestVector)) {
+      throw new IllegalArgumentException("requestVector");
     }
 
-    int nextDot = requestId.indexOf('.');
-    final String coreRequestId = nextDot >= 0 ? requestId.substring(0, nextDot) : requestId;
-    if (coreRequestId.isEmpty()) {
-      throw new IllegalArgumentException("requestId");
+    int nextDot = requestVector.indexOf('.');
+    final String coreRequestVector = nextDot >= 0 ? requestVector.substring(0, nextDot) : requestVector;
+    if (coreRequestVector.isEmpty()) {
+      throw new IllegalArgumentException("requestVector");
     }
 
     final List<Long> suffixes = new ArrayList<>();
     while (nextDot > 0) {
       final int start = nextDot + 1;
-      nextDot = requestId.indexOf('.', start);
-      final String suffixStr = nextDot > 0 ? requestId.substring(start, nextDot) : requestId.substring(start);
+      nextDot = requestVector.indexOf('.', start);
+      final String suffixStr = nextDot > 0 ? requestVector.substring(start, nextDot) : requestVector.substring(start);
       if (suffixStr.isEmpty()) {
-        throw new IllegalArgumentException("requestId");
+        throw new IllegalArgumentException("requestVector");
       }
 
       suffixes.add(Long.parseLong(suffixStr, 16));
     }
 
-    return new RequestIdParts(coreRequestId, suffixes);
+    return new RequestVectorParts(coreRequestVector, suffixes);
   }
 
-  public static final class RequestIdParts {
-    private final String requestId;
+  public static final class RequestVectorParts {
+    private final String requestVector;
     private final List<Long> suffixes;
 
-    public RequestIdParts(String requestId, List<Long> suffixes) {
-      this.requestId = requestId;
+    public RequestVectorParts(String requestVector, List<Long> suffixes) {
+      this.requestVector = requestVector;
       this.suffixes = Collections.unmodifiableList(Arrays.asList(suffixes.toArray(new Long[suffixes.size()])));
     }
 
-    public String getRequestId() {
-      return requestId;
+    public String getRequestVector() {
+      return requestVector;
     }
 
     public List<Long> getSuffixes() {
@@ -116,25 +116,25 @@ public final class TrackingHttpHeaderNames {
     @Override
     public boolean equals(Object o) {
       if (this == o) return true;
-      if (!(o instanceof RequestIdParts)) return false;
+      if (!(o instanceof RequestVectorParts)) return false;
 
-      RequestIdParts that = (RequestIdParts) o;
+      RequestVectorParts that = (RequestVectorParts) o;
 
-      return getRequestId().equals(that.getRequestId()) && getSuffixes().equals(that.getSuffixes());
+      return getRequestVector().equals(that.getRequestVector()) && getSuffixes().equals(that.getSuffixes());
 
     }
 
     @Override
     public int hashCode() {
-      int result = getRequestId().hashCode();
+      int result = getRequestVector().hashCode();
       result = 31 * result + getSuffixes().hashCode();
       return result;
     }
 
     @Override
     public String toString() {
-      final StringBuilder builder = new StringBuilder(requestId.length() + suffixes.size() * 17);
-      builder.append(getRequestId());
+      final StringBuilder builder = new StringBuilder(requestVector.length() + suffixes.size() * 17);
+      builder.append(getRequestVector());
       for (final Long suffix : suffixes) {
         builder.append('.').append(Long.toHexString(suffix));
       }