+++ brikar-common/src/test/java/com/truward/brikar/common/test/tracking/TrackingHttpHeaderNamesTest.java
@@ -20,9 +20,9 @@ public final class TrackingHttpHeaderNamesTest {
   public void shouldParseManuallyConstructedNestedRid() {
     final String nestedRid = "rid.0.1.2.3.4.5.6.7";
 
-    final TrackingHttpHeaderNames.RequestIdParts parts = TrackingHttpHeaderNames.parseRequestId(nestedRid);
+    final TrackingHttpHeaderNames.RequestVectorParts parts = TrackingHttpHeaderNames.parseRequestVector(nestedRid);
 
-    assertEquals("rid", parts.getRequestId());
+    assertEquals("rid", parts.getRequestVector());
     assertEquals(Arrays.asList(0L, 1L, 2L, 3L, 4L, 5L, 6L, 7L), parts.getSuffixes());
     assertEquals(nestedRid, parts.toString());
   }
@@ -31,69 +31,69 @@ public final class TrackingHttpHeaderNamesTest {
   public void shouldParseMaxValueSuffix() {
     final String rid = "rid";
 
-    final String nestedRid = TrackingHttpHeaderNames.getNestedRequestId(rid, Long.MAX_VALUE);
-    final TrackingHttpHeaderNames.RequestIdParts parts = TrackingHttpHeaderNames.parseRequestId(nestedRid);
+    final String nestedRid = TrackingHttpHeaderNames.getNestedRequestVector(rid, Long.MAX_VALUE);
+    final TrackingHttpHeaderNames.RequestVectorParts parts = TrackingHttpHeaderNames.parseRequestVector(nestedRid);
 
-    assertEquals(rid, parts.getRequestId());
+    assertEquals(rid, parts.getRequestVector());
     assertEquals(Collections.singletonList(Long.MAX_VALUE), parts.getSuffixes());
   }
 
   @Test
-  public void shouldGenerateNestedRequestId() {
+  public void shouldGenerateNestedRequestVector() {
     final String rid = "req";
 
-    final String nestedRid = TrackingHttpHeaderNames.getNestedRequestId(rid);
+    final String nestedRid = TrackingHttpHeaderNames.getNestedRequestVector(rid);
 
     assertNotSame(rid, nestedRid);
-    final TrackingHttpHeaderNames.RequestIdParts parts = TrackingHttpHeaderNames.parseRequestId(nestedRid);
-    assertEquals(rid, parts.getRequestId());
+    final TrackingHttpHeaderNames.RequestVectorParts parts = TrackingHttpHeaderNames.parseRequestVector(nestedRid);
+    assertEquals(rid, parts.getRequestVector());
     assertEquals(1, parts.getSuffixes().size());
   }
 
   @Test
-  public void shouldConvertSingleRequestId() {
+  public void shouldConvertSingleRequestVector() {
     final String rid = "r";
 
-    final TrackingHttpHeaderNames.RequestIdParts parts = TrackingHttpHeaderNames.parseRequestId(rid);
+    final TrackingHttpHeaderNames.RequestVectorParts parts = TrackingHttpHeaderNames.parseRequestVector(rid);
 
-    assertEquals(rid, parts.getRequestId());
+    assertEquals(rid, parts.getRequestVector());
     assertEquals(Collections.emptyList(), parts.getSuffixes());
   }
 
   @Test
-  public void shouldGenerateMultipleNestedRequestIds() {
-    final String rid = "testRequestId";
+  public void shouldGenerateMultipleNestedRequestVectors() {
+    final String rid = "testRequestVector";
 
     String nestedRid = rid;
     final List<Long> suffixes = new ArrayList<>();
     final ThreadLocalRandom random = ThreadLocalRandom.current();
     for (int i = 0; i < 100; ++i) {
       long suffix = random.nextLong(0, Long.MAX_VALUE);
-      nestedRid = TrackingHttpHeaderNames.getNestedRequestId(nestedRid, suffix);
+      nestedRid = TrackingHttpHeaderNames.getNestedRequestVector(nestedRid, suffix);
       suffixes.add(suffix);
     }
 
-    final TrackingHttpHeaderNames.RequestIdParts parts = TrackingHttpHeaderNames.parseRequestId(nestedRid);
-    assertEquals(new TrackingHttpHeaderNames.RequestIdParts(rid, suffixes), parts);
+    final TrackingHttpHeaderNames.RequestVectorParts parts = TrackingHttpHeaderNames.parseRequestVector(nestedRid);
+    assertEquals(new TrackingHttpHeaderNames.RequestVectorParts(rid, suffixes), parts);
   }
 
   @Test(expected = IllegalArgumentException.class)
-  public void shouldRejectEmptyRequestIdNesting() {
-    TrackingHttpHeaderNames.getNestedRequestId("", 1L);
+  public void shouldRejectEmptyRequestVectorNesting() {
+    TrackingHttpHeaderNames.getNestedRequestVector("", 1L);
   }
 
   @Test(expected = IllegalArgumentException.class)
-  public void shouldRejectMalformedRequestIdParsingWithTrailingDot() {
-    TrackingHttpHeaderNames.parseRequestId("r.1.");
+  public void shouldRejectMalformedRequestVectorParsingWithTrailingDot() {
+    TrackingHttpHeaderNames.parseRequestVector("r.1.");
   }
 
   @Test(expected = IllegalArgumentException.class)
-  public void shouldRejectMalformedRequestIdParsingWithNeighbourDots() {
-    TrackingHttpHeaderNames.parseRequestId("r.1..2");
+  public void shouldRejectMalformedRequestVectorParsingWithNeighbourDots() {
+    TrackingHttpHeaderNames.parseRequestVector("r.1..2");
   }
 
   @Test(expected = IllegalArgumentException.class)
-  public void shouldRejectMalformedRequestIdParsingWithEmptyCore() {
-    TrackingHttpHeaderNames.parseRequestId(".1.2");
+  public void shouldRejectMalformedRequestVectorParsingWithEmptyCore() {
+    TrackingHttpHeaderNames.parseRequestVector(".1.2");
   }
 }